refactor: move keycloak provides into auth folder
